+++ Braintree/Test/Nonce.cs
@@ -15,5 +15,32 @@ namespace Braintree.Test
         public const string AbstractTransactable = "fake-abstract-transactable-nonce";
         public const string Europe = "fake-europe-bank-account-nonce";
         public const string Coinbase = "fake-coinbase-nonce";
+        public const string TransactableVisa = "fake-valid-visa-nonce";
+        public const string TransactableAmEx = "fake-valid-amex-nonce";
+        public const string TransactableMasterCard = "fake-valid-mastercard-nonce";
+        public const string TransactableDiscover = "fake-valid-discover-nonce";
+        public const string TransactableJCB = "fake-valid-jcb-nonce";
+        public const string TransactableMaestro = "fake-valid-maestro-nonce";
+        public const string TransactableDinersClub = "fake-valid-dinersclub-nonce";
+        public const string TransactablePrepaid = "fake-valid-prepaid-nonce";
+        public const string TransactableCommercial = "fake-valid-commercial-nonce";
+        public const string TransactableDurbinRegulated = "fake-valid-durbin-regulated-nonce";
+        public const string TransactableHealthcare = "fake-valid-healthcare-nonce";
+        public const string TransactableDebit = "fake-valid-debit-nonce";
+        public const string TransactablePayroll = "fake-valid-payroll-nonce";
+        public const string TransactableNoIndicators = "fake-valid-no-indicators-nonce";
+        public const string TransactableUnknownIndicators = "fake-valid-unknown-indicators-nonce";
+        public const string TransactableCountryOfIssuanceUSA = "fake-valid-country-of-issuance-usa-nonce";
+        public const string TransactableCountryOfIssuanceCAD = "fake-valid-country-of-issuance-cad-nonce";
+        public const string TransactableIssuingBankNetworkOnly = "fake-valid-issuing-bank-network-only-nonce";
+        public const string ProcessorDeclinedVisa = "fake-processor-declined-visa-nonce";
+        public const string ProcessorDeclinedMasterCard = "fake-processor-declined-mastercard-nonce";
+        public const string ProcessorDeclinedAmEx = "fake-processor-declined-amex-nonce";
+        public const string ProcessorDeclinedDiscover = "fake-processor-declined-discover-nonce";
+        public const string ProcessorFailureJCB = "fake-processor-failure-jcb-nonce";
+        public const string LuhnInvalid = "fake-luhn-invalid-nonce";
+        public const string PayPalFuturePaymentRefreshToken = "fake-paypal-future-refresh-token-nonce";
+        public const string SEPA = "fake-sepa-bank-account-nonce";
+        public const string GatewayRejectedFraud = "fake-gateway-rejected-fraud-nonce";
     }
 }